refactor: 카드 추가 UI 업데이트 옵저버 패턴으로 분리
